requirements.txt
requests>=2.28.0
xlrd>=2.0.1
gspread>=6.0.0
google-auth>=2.0.0
